Create librarything.cs
Please feel free to rename this as you like because it currently has a dumb name. Also code was totally borrowed from this place: https://msdn.microsoft.com/en-us/library/ezwyzy7b.aspx
and put here for ease of sharing and as a base. Cool. Also make sure to do the right things in vs2013 so that it works before actually trying to do other things.